pedo repariert

+++ Values/Values.h
@@ -67,13 +67,13 @@ class Values {
 
 		// Ram Storage arrays
 		uint16_t co2[CO2_STORAGE_SIZE] = {0};  // in ppm. in flash memory, its divided by 100 to fit in a byte
-		uint8_t co2_idx;
+		uint16_t co2_idx;
 		uint16_t voc[VOC_STORAGE_SIZE] = {0}; // in ppb, bzw. mg/m3
-		uint8_t voc_idx;
+		uint16_t voc_idx;
 		uint8_t uvi[UVI_STORAGE_SIZE] = {0};   // UV index
-		uint8_t uvi_idx;
+		uint16_t uvi_idx;
 		float temp[TEMP_STORAGE_SIZE] = {0}; // in degrees
-		uint8_t temp_idx;
+		uint16_t temp_idx;
 		uint16_t steps;
 
 		bool pedoEnable;
